Support empty version to build AMIs with latest version

+++ packer/gcp/init.sh
@@ -10,6 +10,7 @@ SOURCEGRAPH_DEPLOY_REPO_URL='https://github.com/sourcegraph/deploy.git'
 DEPLOY_PATH='/root/deploy/install'
 USER_ROOT_PATH="/home/sourcegraph"
 SHELL=/bin/bash
+
 ###############################################################################
 # Prepare the system
 ###############################################################################
@@ -56,15 +57,12 @@ echo '@reboot sleep 5 && bash /home/sourcegraph/install.sh' | crontab -
 
 # Add standard bash aliases
 echo "export KUBECONFIG='/etc/rancher/k3s/k3s.yaml'" | tee -a "$USER_ROOT_PATH"/.bash_profile
-echo "export INSTANCE_VERSION='$SOURCEGRAPH_VERSION'" | tee -a /home/sourcegraph/.bash_profile
 echo "export INSTANCE_SIZE='$INSTANCE_SIZE'" | tee -a /home/sourcegraph/.bash_profile
 echo "export SHELL='/bin/bash'" | tee -a "$USER_ROOT_PATH"/.bash_profile
 echo "alias h='helm --kubeconfig /etc/rancher/k3s/k3s.yaml'" | tee -a "$USER_ROOT_PATH"/.bash_profile
 echo "alias k='kubectl --kubeconfig /etc/rancher/k3s/k3s.yaml'" | tee -a "$USER_ROOT_PATH"/.bash_profile
 echo "alias sgupgrade='helm --kubeconfig /etc/rancher/k3s/k3s.yaml upgrade -i -f /home/sourcegraph/deploy/install/override.yaml sourcegraph sourcegraph/sourcegraph'" | tee -a "$USER_ROOT_PATH"/.bash_profile
-
-# Generate files to save instance info in volumes for upgrade purpose
-[ "$SOURCEGRAPH_VERSION" != "" ] && echo "$SOURCEGRAPH_VERSION" | sudo tee "$USER_ROOT_PATH"/.sourcegraph-version
+echo "alias sgrestart='kubectl --kubeconfig /etc/rancher/k3s/k3s.yaml rollout restart deployment sourcegraph-frontend'" | tee -a "$USER_ROOT_PATH"/.bash_profile
 
 ###############################################################################
 # Set up Sourcegraph using Helm
@@ -76,5 +74,12 @@ helm version --short
 # Store Sourcegraph Helm charts locally
 helm repo add sourcegraph https://helm.sourcegraph.com/release
 helm pull --version "$SOURCEGRAPH_VERSION" sourcegraph/sourcegraph
-sleep 5
-sudo mv -f sourcegraph-"$SOURCEGRAPH_VERSION".tgz sourcegraph-charts.tgz
+
+# Generate files to save instance info in volumes for upgrade purpose
+LATEST_VERSION=$(/usr/local/bin/helm search repo sourcegraph -o yaml --kubeconfig /etc/rancher/k3s/k3s.yaml | grep 'app_version' | head -1 | cut -d ":" -f 2 | xargs)
+[ "$SOURCEGRAPH_VERSION" == "" ] && [ "$LATEST_VERSION" != "" ] && SOURCEGRAPH_VERSION=$LATEST_VERSION
+if [ "$SOURCEGRAPH_VERSION" != "" ]; then
+    echo "$SOURCEGRAPH_VERSION" | sudo tee "$USER_ROOT_PATH"/.sourcegraph-version
+    echo "export INSTANCE_VERSION='$SOURCEGRAPH_VERSION'" | tee -a /home/sourcegraph/.bash_profile
+    [ -f sourcegraph-"$SOURCEGRAPH_VERSION".tgz ] && sudo mv -f sourcegraph-"$SOURCEGRAPH_VERSION".tgz sourcegraph-charts.tgz
+fi
